move test into own
